requirements.txt
pip==23.2
cupy-cuda11x==13.0.0
einops==0.7.0
ema-pytorch==0.4.2
huggingface-hub==0.20.3
imageio==2.34.0
lpips==0.1.4
multiprocess==0.70.16
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.9.0.80
pillow==9.5.0
pytorch-fid==0.3.0
pytorch-lightning==1.7.7
requests==2.28.1
scikit-image==0.22.0
scipy==1.11.1
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.16.2
timm==0.9.16
torchmetrics==0.11.4
torchsummary==1.5.1
tqdm==4.66.2
